+++ NEWS.md
@@ -5,8 +5,11 @@ Development version
 
 New arguments for `get_route()` 
 
+* `departure_time` 
 * `alternatives` argument for displaying alternative routes
 * `avoid` - tolls, highways, ferries, indoors
+* `traffic_model` - best_guess, optimistic, pessimistic
+* `units` - metric or imperial (only the displayed text - value field is always metric)
 
 ## Version 0.2.0